harden and rename electron sqlite rebuild script
